refactor: move AppScenarios interface into apptests module
