+++ apptests/appscenarios/appscenarios.go
@@ -18,14 +18,6 @@ import (
 
 var upgradeKAppsRepoPath string
 
-// AppScenario defines the behavior and name of an application test scenario
-type AppScenario interface {
-	Name() string                                    // scenario name
-	Install(context.Context, *environment.Env) error // logic implemented by a scenario
-	InstallPreviousVersion(ctx context.Context, env *environment.Env) error
-	Upgrade(ctx context.Context, env *environment.Env) error
-}
-
 // absolutePathTo returns the absolute path to the given application directory.
 func absolutePathTo(application string) (string, error) {
 	wd, err := os.Getwd()
